Squashed commit of the following:
commit 44138933b299fff9cf044dc825216d73e82dbfe3
Author: chema <chema@localhost>
Date:   Thu Jan 16 17:17:27 2020 +0100

	DATACASS-718: Integer Overflow using Update

+++ spring-data-cassandra/src/main/java/org/springframework/data/cassandra/core/query/Update.java
@@ -662,7 +662,7 @@ public class Update {
 		@Override
 		public String toString() {
 			return String.format("%s = %s %s %d", getColumnName(), getColumnName(), value.doubleValue() > 0 ? "+" : "-",
-					Math.abs(value.intValue()));
+					Math.abs(value.longValue()));
 		}
 	}
